lang/codes/banagm/bnms/requirements.txt
aiomysql==0.0.22
anyio==3.3.4
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
contextvars==2.4
crypto==1.4.1
cryptography==35.0.0
databases==0.5.3
dataclasses==0.6
ecdsa==0.17.0
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
immutables==0.16
Naked==0.1.31
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyMySQL==0.9.3
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
shellescape==3.8.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.26
starlette==0.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
